test: migrate resource test unit to rspec
