+++ .gitignore
@@ -4,3 +4,4 @@ log/*.log
 tmp/**/*
 .idea
 .rspec
+vendor/bundle
